Create separate classes for pde-mgmt HTTP handling

+++ plantdescriptionengine/src/main/java/eu/arrowhead/core/plantdescriptionengine/services/orchestration_mgmt/OrchestratorClient.java
@@ -217,6 +217,11 @@ public class OrchestratorClient implements PlantDescriptionUpdateListener {
         }
 
         List<Integer> rulesToRemove = ruleMap.get(entry.id());
+
+        if (rulesToRemove == null) {
+            return Future.done();
+        }
+
         var deletions = rulesToRemove.stream().map(ruleId -> deleteRule(ruleId)).collect(Collectors.toList());
 
         return Futures.serialize(deletions).flatMap(result -> {
